+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.10] - 2023-08-16
+### Added
+* documentation for confusing code snippets within `AStarPathfinding`
+### Changed
+* `AStarPathfinding`'s `TryTrimPath()` to `TryReduceOutsideMovementRangePath()`
+* `IPathfinding`'s `CanRepeatNodesInPath` to `CanCreatePathsWithRepeatNodes`
+* `IPathfinding`'s `IsPathTrimmable()` to `IsPathOutsideMovementRange()`
+### Fixed
+* bug where an `OutsideMovementRange` path would be recreated if user removed penultimate cell
+
 ## [0.0.9] - 2023-08-01
 ### Added
 * `CanRepeatNodesInPath` to `IPathfinding` to allow for repeating nodes or retracing nodes when calling `TryAddNodeToPath`